Fix #66: read driver options through a typed DriverConfig value object

+++ config/aether.php
@@ -102,10 +102,9 @@ return [
 
         'local' => [
             'synchronous_safe' => true,
-
             // The number of qubits (random bits) produced per shot of the
             // entropy circuit. Must fit within max_qubits below.
-            'entropy_qubits' => (int) env('AETHER_ENTROPY_QUBITS', 16),
+            'entropy_qubits' => env('AETHER_ENTROPY_QUBITS', 16),
 
             // The local simulator has no real task queue: a dispatched circuit
             // runs immediately and its result is cached under a synthetic task
@@ -118,8 +117,10 @@ return [
             // use doubles with every additional qubit. The default of 25
             // caps that at 2^25 x 16 bytes ~= 512 MB. Raise it only once
             // you've confirmed the host has memory to spare, or set it to
-            // null to remove the ceiling entirely. Applies to ->run(),
-            // ->dispatch(), Quantum::batch() and entropy generation alike.
+            // null to remove the ceiling entirely. A positive integer, or
+            // null/blank for no ceiling; anything else throws
+            // InvalidDriverConfigException when the driver is resolved. Applies
+            // to ->run(), ->dispatch(), Quantum::batch() and entropy generation alike.
             'max_qubits' => env('AETHER_MAX_QUBITS', 25),
         ],
 
@@ -131,12 +132,11 @@ return [
             'bucket' => env('AETHER_S3_BUCKET'),
             'device_arn' => env('AETHER_DEVICE_ARN', 'arn:aws:braket:::device/quantum-simulator/amazon/sv1'),
             'synchronous_safe' => true,
-
             // The number of qubits (random bits) produced per shot of the
             // entropy circuit. Must fit within max_qubits below; each
             // generate() call is one task whose estimated cost is checked
             // against max_cost_per_run.
-            'entropy_qubits' => (int) env('AETHER_ENTROPY_QUBITS', 16),
+            'entropy_qubits' => env('AETHER_ENTROPY_QUBITS', 16),
 
             // No ceiling here: Braket enforces its own per-device qubit
             // limits, so this package does not duplicate or guess at those.
@@ -148,10 +148,12 @@ return [
             // (e.g. SV1) bill per-minute instead, but the task+shot model
             // is what estimateCost() covers; treat simulator estimates as
             // a rough proxy, not an exact figure. Override via env/config
-            // without a package release.
+            // without a package release. The rates are strictly validated (not silently cast to 0)
+            // when the driver is resolved: a non-numeric or negative value
+            // throws InvalidDriverConfigException instead of pricing at 0.
             'pricing' => [
-                'per_task' => (float) env('AETHER_AWS_PRICE_PER_TASK', 0.30),
-                'per_shot' => (float) env('AETHER_AWS_PRICE_PER_SHOT', 0.00035),
+                'per_task' => env('AETHER_AWS_PRICE_PER_TASK', 0.30),
+                'per_shot' => env('AETHER_AWS_PRICE_PER_SHOT', 0.00035),
                 'currency' => env('AETHER_AWS_PRICE_CURRENCY', 'USD'),
             ],
 
@@ -160,7 +162,9 @@ return [
             // amount — for a batch, the total of all its circuits — before
             // any AWS call is made. null (default) means unlimited. Requires
             // the pricing rates above: a ceiling with no rates fails fast
-            // instead of never tripping.
+            // instead of never tripping. A non-negative number, or null/blank
+            // for no ceiling; anything else throws InvalidDriverConfigException
+            // when the driver is resolved.
             'max_cost_per_run' => env('AETHER_AWS_MAX_COST'),
         ],
